refactor: Rename to useLogin to match JavaScript naming convention
